// gis/types/Student.ts
// types.ts
export type Student = {
  id: string;
  level: number;
  teritorija: string;
  starost: string;
  spol: string;
  ukupno: string;
  neSkolujeSe: string;
  predskolskoObrazovanje: string;
  osnovnaSkola: string;
  srednjaSkola: string;
  specijalizacijaPoslijeSrednje: string;
  visaSkola: string;
  stariProgramOsnovne: string;
  stariProgramSpecijalisticke: string;
  stariProgramMagistarske: string;
  stariProgramDoktorske: string;
  programBolonjaI: string;
  programBolonjaII: string;
  programBolonjaIntegrisani: string;
  programBolonjaIII: string;
};